welcome
In this code, we start with the <!DOCTYPE html> declaration, which tells the browser that this is an HTML document. The <html> element is the root element of the webpage.

Inside the <html> element, we have the <head> element that contains meta-information about the document. In this example, we have a <title> element that sets the title of the webpage, which appears in the browser's title bar or tab.